+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added support for private DNS for Amazon Virtual Private Clouds, health check
   failure reasons, and reusable delegation sets to the Amazon Route 53 client.
+* Updated the CloudFront model.
+
 
 ## 2.7.2 - 2014-10-23